Rename file names and prepare for graphql sources
